+++ src/main/java/duke/Duke.java
@@ -1,3 +1,13 @@
+package duke;
+
+import duke.commands.Command;
+import duke.commands.CommandResult;
+import duke.exceptions.DukeException;
+import duke.task.TaskList;
+import duke.parser.Parser;
+import duke.storage.Storage;
+import duke.ui.Ui;
+
 import java.io.IOException;
 import java.util.Scanner;
