+++ src/main/java/duke/commands/ExitCommand.java
@@ -1,3 +1,5 @@
+package duke.commands;
+
 public class ExitCommand extends Command {
     public static final String COMMAND_WORD = "bye";
