+++ src/main/java/duke/parser/Parser.java
@@ -1,3 +1,16 @@
+package duke.parser;
+
+import duke.commands.Command;
+import duke.commands.ExitCommand;
+import duke.commands.ListCommand;
+import duke.commands.MarkCommand;
+import duke.commands.UnmarkCommand;
+import duke.commands.DeleteCommand;
+import duke.commands.TodoCommand;
+import duke.commands.DeadlineCommand;
+import duke.commands.EventCommand;
+import duke.exceptions.DukeException;
+
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
