+++ src/main/java/duke/storage/Storage.java
@@ -1,3 +1,12 @@
+package duke.storage;
+
+import duke.exceptions.DukeException;
+import duke.task.TaskList;
+import duke.task.Task;
+import duke.task.Todo;
+import duke.task.Deadline;
+import duke.task.Event;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
